# specs/001-quickflow-productivity/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: QuickFlow API
  version: 1.0.0
  description: >
    REST API of QuickFlow (single-user personal productivity). Source of truth for the backend springdoc swagger
    and the generated typescript-angular client. Errors use RFC 9457 application/problem+json.
servers:
- {url: 'http://localhost:8080'}
...
paths:
  /api/tasks:
    get:
      tags: [Tasks]
      operationId: listTasks
      summary: List tasks (non-archived by default) with search, filters and sorting
      parameters:
      - name: q
        in: query
        required: false
        schema: {type: string}
        description: Case-insensitive title contains
      - name: status
        in: query
        required: false
        schema: {$ref: '#/components/schemas/TaskStatus'}
      - name: priority
        in: query
        required: false
        schema: {$ref: '#/components/schemas/TaskPriority'}
      - name: dueFrom
        in: query
        required: false
        schema: {type: string, format: date}
      - name: dueTo
        in: query
        required: false
        schema: {type: string, format: date}
      - name: overdue
        in: query
        required: false
        schema: {type: boolean}
        description: true = only overdue tasks
      - name: archived
        in: query
        required: false
        schema: {type: boolean, default: false}
        description: true = only archived tasks
      - name: sort
        in: query
        required: false
        schema: {$ref: '#/components/schemas/TaskSort'}
      - name: direction
        in: query
        required: false
        schema: {$ref: '#/components/schemas/SortDirection'}
      responses:
        '200':
          description: Matching tasks (tasks without due date sort last when sort=DUE_DATE)
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Task'}
        '400': {$ref: '#/components/responses/BadRequest'}
    post:
      tags: [Tasks]
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TaskRequest'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Task'}
        '400': {$ref: '#/components/responses/BadRequest'}
  /api/tasks/{id}:
    parameters:
    - name: id
      in: path
      required: true
      schema: {type: integer, format: int64}
    get:
      tags: [Tasks]
      operationId: getTask
      responses:
        '200':
          description: The task
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Task'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      tags: [Tasks]
      operationId: updateTask
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TaskRequest'}
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Task'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [Tasks]
      operationId: deleteTask
      responses:
        '204': {description: Deleted}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/tasks/{id}/complete:
    parameters:
    - name: id
      in: path
      required: true
      schema: {type: integer, format: int64}
    post:
      tags: [Tasks]
      operationId: completeTask
      summary: Set status DONE and completedAt
      responses:
        '200':
          description: Completed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Task'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/tasks/{id}/archive:
    parameters:
    - name: id
      in: path
      required: true
      schema: {type: integer, format: int64}
    post:
      tags: [Tasks]
      operationId: archiveTask
      responses:
        '200':
          description: Archived
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Task'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/tasks/{id}/restore:
    parameters:
    - name: id
      in: path
      required: true
      schema: {type: integer, format: int64}
    post:
      tags: [Tasks]
      operationId: restoreTask
      responses:
        '200':
          description: Restored
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Task'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  responses:
    BadRequest:
      description: Validation error
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details
      properties:
        type: {type: string}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string}
        errors:
          type: array
          items: {$ref: '#/components/schemas/FieldError'}
    FieldError:
      type: object
      required: [field, message]
      properties:
        field: {type: string}
        message: {type: string}
    TaskStatus:
      type: string
      enum: [TODO, IN_PROGRESS, DONE]
    TaskPriority:
      type: string
      enum: [LOW, MEDIUM, HIGH]
    TaskSort:
      type: string
      enum: [DUE_DATE, CREATED_AT]
      default: CREATED_AT
    SortDirection:
      type: string
      enum: [ASC, DESC]
      default: DESC
    TaskRequest:
      type: object
      required: [title]
      properties:
        title: {type: string, minLength: 1, maxLength: 200}
        description:
          type: [string, 'null']
          maxLength: 2000
        status: {$ref: '#/components/schemas/TaskStatus'}
        priority: {$ref: '#/components/schemas/TaskPriority'}
        dueDate:
          type: [string, 'null']
          format: date
    Task:
      type: object
      required: [id, title, status, priority, createdAt, updatedAt, archived, overdue]
      properties:
        id: {type: integer, format: int64}
        title: {type: string}
        description:
          type: [string, 'null']
        status: {$ref: '#/components/schemas/TaskStatus'}
        priority: {$ref: '#/components/schemas/TaskPriority'}
        dueDate:
          type: [string, 'null']
          format: date
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        completedAt:
          type: [string, 'null']
          format: date-time
        archived: {type: boolean}
        overdue: {type: boolean}
...